Move apple_tv base entity to separate module
