+++ homeassistant/components/apple_tv/const.py
@@ -6,3 +6,6 @@ CONF_CREDENTIALS = "credentials"
 CONF_IDENTIFIERS = "identifiers"
 
 CONF_START_OFF = "start_off"
+
+SIGNAL_CONNECTED = "apple_tv_connected"
+SIGNAL_DISCONNECTED = "apple_tv_disconnected"
